# Improve polygons