--- a/examples/polygons/index.jade
+++ b/examples/polygons/index.jade
@@ -1,0 +1,1 @@
+extends ../common/templates/index.jade