--- a/src/polygonFeature.js
+++ b/src/polygonFeature.js
@@ -8,6 +8,24 @@ var feature = require('./feature');
  *
  * @class geo.polygonFeature
  * @extends geo.feature
+ * @param {Object} arg Options object
+ * @param {Object|Function} [arg.position] Position of the data.  Default is
+ *   (data).  The position is an Object which specifies the location of the
+ *   data in geo-spatial context.
+ * @param {Object|Function} [arg.polygon] Polygons from the data.  Default is
+ *   (data).  Typically, the data is an array of polygons, each of which is
+ *   of the form {outer: [(coordinates)], inner: [[(coordinates of first
+ *   hole)], [(coordinates of second hole)], ...]}.  The inner record is
+ *   optional.  Alternately, if there are no holes, a polygon can just be an
+ *   array of coordinates.  Coordinates are in the form {x: (x), y: (y),
+ *   z: (z)}, with z being optional.  The first and last point of each polygon
+ *   must be the same.
+ * @param {Object} [arg.style] Style object with default style options.
+ * @param {Object|Function} [arg.style.fillColor] Color to fill each polygon.
+ *   The color can vary by vertex.  Colors can be css names or hex values, or
+ *   an object with r, g, b on a [0-1] scale.
+ * @param {Object|Function} [arg.style.fillOpacity] Opacity for each polygon.
+ *   The opacity can vary by vertex.  Opacity is on a [0-1] scale.
  * @returns {geo.polygonFeature}
  */
 //////////////////////////////////////////////////////////////////////////////
@@ -51,8 +69,12 @@ var polygonFeature = function (arg) {
 
   ////////////////////////////////////////////////////////////////////////////
   /**
-   * Override the parent data method to keep track of changes to the
-   * internal coordinates.
+   * Get/set data.
+   *
+   * @memberof geo.polygonFeature
+   * @param {Object} [data] if specified, use this for the data and return the
+   *    feature.  If not specified, return the current data.
+   * @returns {geo.polygonFeature|Object}
    */
   ////////////////////////////////////////////////////////////////////////////
   this.data = function (arg) {
@@ -68,6 +90,7 @@ var polygonFeature = function (arg) {
    * Get the internal coordinates whenever the data changes.  For now, we do
    * the computation in world coordinates, but we will need to work in GCS
    * for other projections.
+   * @memberof geo.polygonFeature
    * @private
    */
   ////////////////////////////////////////////////////////////////////////////
@@ -78,7 +101,7 @@ var polygonFeature = function (arg) {
       var poly = polyFunc(d);
       var outer, inner;
 
-      outer = (poly.outer || []).map(function (d0, j) {
+      outer = (poly.outer || (poly instanceof Array ? poly : [])).map(function (d0, j) {
         return posFunc.call(m_this, d0, j, d, i);
       });
 
@@ -96,9 +119,12 @@ var polygonFeature = function (arg) {
 
   ////////////////////////////////////////////////////////////////////////////
   /**
-   * Get/Set polygon accessor
+   * Get/set polygon accessor.
    *
-   * @returns {geo.pointFeature}
+   * @memberof geo.polygonFeature
+   * @param {Object} [polygon] if specified, use this for the polygon accesor
+   *    and return the feature.  If not specified, return the current polygon.
+   * @returns {geo.polygonFeature|Object}
    */
   ////////////////////////////////////////////////////////////////////////////
   this.polygon = function (val) {
@@ -115,9 +141,12 @@ var polygonFeature = function (arg) {
 
   ////////////////////////////////////////////////////////////////////////////
   /**
-   * Get/Set position accessor
+   * Get/Set position accessor.
    *
-   * @returns {geo.pointFeature}
+   * @memberof geo.polygonFeature
+   * @param {Object} [position] if specified, use this for the position accesor
+   *    and return the feature.  If not specified, return the current position.
+   * @returns {geo.polygonFeature|Object}
    */
   ////////////////////////////////////////////////////////////////////////////
   this.position = function (val) {
@@ -134,8 +163,10 @@ var polygonFeature = function (arg) {
 
   ////////////////////////////////////////////////////////////////////////////
   /**
-   * Point searce method for selection api.  Returns markers containing the
+   * Point search method for selection api.  Returns markers containing the
    * given point.
+   *
+   * @memberof geo.polygonFeature
    * @argument {object} coordinate
    * @returns {object}
    */
@@ -162,6 +193,7 @@ var polygonFeature = function (arg) {
   ////////////////////////////////////////////////////////////////////////////
   /**
    * Initialize
+   * @memberof geo.polygonFeature
    */
   ////////////////////////////////////////////////////////////////////////////
   this._init = function (arg) {