--- a/testing/test-data/land_polygons.json.url
+++ b/testing/test-data/land_polygons.json.url
@@ -1,0 +1,1 @@
+https://data.kitware.com/api/v1/file/5762bd868d777f68be8f4017/download