… examples/polygons/example.json
… examples/polygons/main.js
… examples/polygons/thumb.jpg
… src/util/init.js
… testing/test-data/land_polygons.json.md5
… tests/cases/polygonFeature.js