@manthey 2016-06-17
Some performance comparisons (all times in milliseconds) using world-wide 10 arcminute bathymetric data with 5779 polygons with lots of holes that convert to 2,209,058 triangles (6,627,174 vertices) from a 51 Mb json file.

| Step | Before | After | Comments |
|-------|---------:|-------:|----------------|
| getCoordinates | 1305 | 160 | This is getting called twice based on how the example is run.  This is the slowest of the two times |
| createGlPolygons -- first run | 21405 | 16017 | Earcut takes around 14000 ms.  The rest is geometry and style calculations |
| createGlPolygons -- highlighting | 27280 | 5.39 | The before timing started out taking around 19500 ms, but gradually increased due to garbage collection |
| pointSearch | 101 | 2.26 | |

The speed improvements come from changing map functions to loops, only calculating triangles as needed, calculating bounding boxes of polygons, and taking advantage of uniform values.