@aashish24 2016-06-20
> Some performance comparisons

this is cool. For the point search, the speed up is because of the use of bbox?