@manthey 2016-06-27
It is also changing from the forEach to a loop, and not bothering to check
the holes if the point wasn't inside the outer polygon.

On Mon, Jun 20, 2016 at 10:45 AM, Aashish Chaudhary <
notifications@github.com> wrote:

> Some performance comparisons
>
> this is cool. For the point search, the speed up is because of the use of
> bbox?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/OpenGeoscience/geojs/pull/594#issuecomment-227163602>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/AIX_R5djQMfTNe5QvO24yTDIHBnWCTT-ks5qNqeVgaJpZM4I3hEE>
> .
>



-- 
David Manthey
R&D Engineer
Kitware Inc.
(518) 881-4439